Salesbot/requirements.txt
langchain>=0.0.205
openai>=0.27.8 
python-dotenv==1.0.0
faiss-cpu==1.7.4